finish

+++ controller/AdminController.js
@@ -8,28 +8,39 @@ const PeralatanTbl = require("../model/PeralatanTbl");
 const TipsTbl = require("../model/TipsTbl");
 const RambuTbl = require("../model/RambuTbl");
 const SlideShowTbl = require("../model/SlideShowTbl");
+const { type } = require("os");
+const DashboardModel = require("../model/DashboardModel");
 class AdminController {
+    async getDasboard(req, res) {
+        const data = await DashboardModel.get();
+        res.json({ status: "ok", data })
+    }
+
     async getPengertian(req, res) {
         const data = await new PengertianTbl().getAll();
         res.json({ data: data });
     }
     async savePengertian(req, res) {
         const formData = req.body;
-        const tmp_name = req.file.path;
+        const { status, id } = typeof formData.id_pengertian != "undefined" ? await new PengertianTbl().update(formData.id_pengertian, formData) : await new PengertianTbl().save(formData);
 
-        const { status, id } = await new PengertianTbl().save(formData);
         if (status == "sukses") {
+
             const output = "assets/pengertian/" + id + ".jpg";
             try {
-                const image = await Jimp.read(tmp_name);
-                image.crop(parseInt(formData.x), parseInt(formData.y), parseInt(formData.width), parseInt(formData.height));
-                await image.writeAsync(output);
-                console.log('Gambar berhasil dipotong!');
-                fs.unlink(tmp_name, (err) => {
-                    if (!err) {
-                        console.log("sukses");
-                    }
-                })
+                if (typeof req.file !== "undefined") {
+                    const tmp_name = req.file.path;
+                    const image = await Jimp.read(tmp_name);
+                    image.crop(parseInt(formData.x), parseInt(formData.y), parseInt(formData.width), parseInt(formData.height));
+                    await image.writeAsync(output);
+                    console.log('Gambar berhasil dipotong!');
+                    fs.unlink(tmp_name, (err) => {
+                        if (!err) {
+                            console.log("sukses");
+                        }
+                    })
+                }
+
             } catch (error) {
                 console.error('Kesalahan dalam memotong gambar:', error);
             }
@@ -45,14 +56,16 @@ class AdminController {
         const _hapus = await new PengertianTbl().delete(id)
         const file = "assets/pengertian/" + id + ".jpg";
         if (_hapus == "sukses") {
-            const c = fs.existsSync(file);
-            if (c) {
-                fs.unlink(file, (err) => {
-                    if (!err) {
-                        console.log("berhasil di hapus")
-                    }
-                })
-            }
+            fs.access(file, fs.constants.F_OK, (err) => {
+                if (!err) {
+                    fs.unlink(file, (err) => {
+                        if (!err) {
+                            console.log("berhasil di hapus")
+                        }
+                    })
+                }
+            })
+
             res.json({ status: "deleted" })
         }
         else {
@@ -64,24 +77,34 @@ class AdminController {
         const data = await new ArtikelTbl().getAll();
         res.json({ data: data });
     }
+    async getArtikelSingle(req, res) {
+        const { id } = req.params;
+        const data = await new ArtikelTbl().getSingle(id);
+        res.json({
+            data: data
+        })
+    }
     async save_artikel(req, res) {
         const form_data = req.body;
-        const tmp_name = req.file.path;
-        console.log(tmp_name);
-        const save = await new ArtikelTbl().simpan(form_data);
+
+
+        const save = (typeof form_data.id_artikel != "undefined") ? await new ArtikelTbl().update(form_data.id_artikel, form_data) : await new ArtikelTbl().simpan(form_data);
         if (save.status == "sukses") {
+            if (typeof req.file != "undefined") {
+                const tmp_name = req.file.path;
+                const img = await Jimp.read(tmp_name);
+                img.crop(
+                    parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height)
 
-            const img = await Jimp.read(tmp_name);
-            img.crop(
-                parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height)
+                );
+                await img.writeAsync("assets/artikel/" + save.id + ".jpg");
+                fs.unlink(tmp_name, (err) => {
+                    if (!err) {
+                        console.log("sukses");
+                    }
+                })
 
-            );
-            await img.writeAsync("assets/artikel/" + save.id + ".jpg");
-            fs.unlink(tmp_name, (err) => {
-                if (!err) {
-                    console.log("sukses");
-                }
-            })
+            }
 
             res.json({ status: "data_saved" })
         }
@@ -93,9 +116,11 @@ class AdminController {
         const hapus = await new ArtikelTbl().delete(id_artikel)
         if (hapus.status == "sukses_deleted") {
             const path = `assets/artikel/${id_artikel}.jpg`
-            const c = fs.existsSync(path)
-            if (c)
-                fs.unlinkSync(path)
+            fs.access(path, fs.constants.F_OK, (err) => {
+                if (!err) {
+                    fs.unlinkSync(path)
+                }
+            })
             res.json({ status: "data_terhapus" })
         }
         else {
@@ -107,15 +132,27 @@ class AdminController {
         const data = await new PeralatanTbl().getAll();
         res.json({ data: data })
     }
+    async getPeralatanSingle(req, res) {
+        const { id } = req.params;
+        const data = await new PeralatanTbl().getSingle(id);
+        console.log(data);
+        res.json({ data: data });
+    }
     async simpanPeralatan(req, res) {
         const form_data = req.body;
-        const tmp_file = req.file.path;
-        const simpan = await new PeralatanTbl().simpan(form_data);
+
+
+        const simpan = typeof req.body.id_peralatan !== "undefined" ? await new PeralatanTbl().update(req.body.id_peralatan, form_data) : await new PeralatanTbl().simpan(form_data);
         if (simpan.status == "sukses") {
-            const jimp = await Jimp.read(tmp_file);
-            jimp.crop(parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height))
-            await jimp.writeAsync("assets/peralatan/" + simpan.id + ".jpg");
-            fs.unlinkSync(tmp_file);
+
+            if (typeof req.file !== "undefined") {
+                const tmp_file = req.file.path;
+                const jimp = await Jimp.read(tmp_file);
+                jimp.crop(parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height))
+                await jimp.writeAsync("assets/peralatan/" + simpan.id + ".jpg");
+                fs.unlinkSync(tmp_file);
+            }
+
             res.json({
                 status: "data_saved"
             })
@@ -125,7 +162,6 @@ class AdminController {
                 status: "data_saved"
             })
         }
-
     }
     async deletePeralatan(req, res) {
         const { id_peralatan } = req.params
@@ -148,9 +184,17 @@ class AdminController {
             data: data
         });
     }
+    async getTipsSingle(req, res) {
+        const { id } = req.params;
+        const data = await new TipsTbl().getSingleTips(id)
+        res.json({
+            data: data,
+        })
+    }
     async saveTips(req, res) {
         const form_data = req.body;
-        const simpan = await new TipsTbl().save(form_data)
+
+        const simpan = typeof form_data.id_tips != "undefined" ? await new TipsTbl().update(form_data.id_tips, form_data) : await new TipsTbl().save(form_data)
         if (simpan.status == "sukses") {
             res.json({
                 status: "sukses"
@@ -185,21 +229,33 @@ class AdminController {
     }
     async saveRambu_rambu(req, res) {
         const form_data = req.body;
-        const { status, id } = await new RambuTbl().save(form_data);
-        const tmp_name = req.file.path;
+        const { status, id } = (typeof form_data.id_rambu != "undefined") ? await new RambuTbl().update(form_data.id_rambu, form_data) : await new RambuTbl().save(form_data);
+        console.log(status);
         if (status == "sukses") {
+            if (typeof req.file !== "undefined") {
+                const tmp_name = req.file.path;
 
-            const c = await Jimp.read(tmp_name);
-            c.crop(parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height),);
-            const save = await c.writeAsync("assets/rambu-rambu/" + id + ".png");
-            if (save) {
-                fs.unlinkSync(tmp_name);
+
+                const c = await Jimp.read(tmp_name);
+                c.crop(parseInt(form_data.x), parseInt(form_data.y), parseInt(form_data.width), parseInt(form_data.height),);
+                const save = await c.writeAsync("assets/rambu-rambu/" + id + ".png");
+                if (save) {
+                    fs.unlinkSync(tmp_name);
+                }
             }
+
             res.json({ status: "data_saved" })
         }
         else {
             res.json({ status: "gagal_saved" })
         }
+    }
+    async getRambuSingle(req, res) {
+        const { id } = req.params;
+        const data = await new RambuTbl().getSingle(id);
+        res.json({
+            data: data
+        })
     }
     async deleteRambu_rambu(req, res) {
         const { id } = req.params;
@@ -243,6 +299,11 @@ class AdminController {
         }
 
 
+    }
+    async getPengertianSingle(req, res) {
+        const { id } = req.params;
+        const data = await new PengertianTbl().getSingle(id);
+        res.json({ data: data });
     }
     async deleteSlideShow(req, res) {
         const { id } = req.params;
